+++ rs/ic_os/vsock/src/protocol/mod.rs
@@ -3,3 +3,6 @@ mod utils;
 
 pub use structures::*;
 pub use utils::*;
+
+#[cfg(test)]
+mod tests;
